// src/shared/interface/component/component-tag.interface.ts
/**
 * Type tag component for text
 * @author Irfan Andriansyah <irfan@99.co>
 * @since 2019.10.20
 */
export type ComponentTextTag =
    | 'h1'
    | 'h2'
    | 'h3'
    | 'h4'
    | 'h5'
    | 'h6'
    | 'p'
    | 'span';

/**
 * Type tag component for t
 * @author Irfan Andriansyah <irfan@99.co>
 * @since 2019.10.24
 */
export type ComponentInpuTextTag = 'input' | 'textarea';