// src/shared/interface/component/component-typography.interface.ts
/**
 * Type component for font family type
 * @author Irfan Andriansyah <irfan@99.co>
 * @since 2020.03.13
 */
export type ComponentFontfaceTypography = 'primary' | 'secondary';

/**
 * Type component for color type
 * @author Irfan Andriansyah <irfan@99.co>
 * @since 2020.03.13
 */
export type ComponentColorTypography = '';

/**
 * Type component for styling
 * @author Irfan Andriansyah <irfan@99.co>
 * @since 2020.03.13
 */
export type ComponentStylingTypography =
    | 'heading-1'
    | 'heading-2'
    | 'heading-3'
    | 'heading-4'
    | 'heading-5'
    | 'heading-6'
    | 'default'
    | 'featured'
    | 'meta'
    | 'caption'
    | 'tiny';

/**
 * Type component for font weight type
 * @author Irfan Andriansyah <irfan@99.co>
 * @since 2020.03.13
 */
export type ComponentFontWeightTypography = 300 | 400 | 500 | 600 | 700;